+++ src/publisher/react-integration/index.ts
@@ -2,18 +2,21 @@ import { createIntegrationCore } from "./core";
 import { createReactDevtoolsHookHandlers } from "./devtools-hook-handlers";
 import { createReactInteractionApi } from "./interaction-api";
 import { ReactInternals, ReactIntegration, RecordEventHandler } from "../types";
-// import { dispatcherTrap } from "./dispatcher-trap";
+import { dispatcherTrap } from "./dispatcher-trap";
 
 export function attach(
   renderer: ReactInternals,
   recordEvent: RecordEventHandler
 ): ReactIntegration {
   const integrationCore = createIntegrationCore(renderer);
-
-  // dispatcherTrap(renderer, recordEvent, integrationCore);
+  const dispatcherApi = dispatcherTrap(renderer);
 
   return {
-    ...createReactDevtoolsHookHandlers(integrationCore, recordEvent),
+    ...createReactDevtoolsHookHandlers(
+      integrationCore,
+      dispatcherApi,
+      recordEvent
+    ),
     ...createReactInteractionApi(integrationCore),
   };
 }