+++ src/publisher/react-integration/utils/isPlainObject.ts
@@ -1,0 +1,8 @@
+export function isPlainObject(value: any) {
+  return (
+    value !== null &&
+    typeof value === "object" &&
+    value.constructor === Object &&
+    typeof value.$$typeof !== "symbol"
+  );
+}