+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed crash when used for React prior 16.9 because of another internals for dependencies (context)
 - Improved object diff for update events
+- Added hook call stack for `useState` and `useReducer` hooks
 
 ## 0.3.0 (September 8, 2021)